+++ www/package.json
@@ -52,7 +52,7 @@
     "grommet": "^2.23.0",
     "grommet-icons": "^4.7.0",
     "highlight.js": "^11.5.1",
-    "honorable": "^0.94.0",
+    "honorable": "^0.95.0",
     "honorable-theme-default": "^0.48.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
