+++ www/src/components/Plural.js
@@ -12,7 +12,7 @@ import Terraform from './repos/Terraform'
 import EditUser from './users/EditUser'
 import { IntegrationPage } from './repos/Integrations'
 import Invoices from './payments/Invoices'
-import Sidebar from './Sidebar'
+import Sidebar from './layout/Sidebar'
 import Publishers from './publisher/Publishers'
 import Explore from './explore/Explore'
 import Repository from './repository/Repository'
