+++ www/src/components/Toolbar.js
@@ -5,7 +5,7 @@ import { Box } from 'grommet'
 import Me from './users/Me'
 import SearchRepositories from './repos/SearchRepositories'
 import { CurrentUserContext } from './login/CurrentUser'
-import { SIDEBAR_WIDTH, SMALL_WIDTH } from './Sidebar'
+import { SIDEBAR_WIDTH, SMALL_WIDTH } from './layout/Sidebar'
 import { Breadcrumbs } from './Breadcrumbs'
 import { Notifications } from './users/Notifications'
 import { LoopingLogo } from './utils/AnimatedLogo'
