+++ www/src/components/utils/SectionChoice.js
@@ -1,6 +1,6 @@
 import { Box, Text } from 'grommet'
 
-import { SIDEBAR_ICON_HEIGHT } from '../Sidebar'
+import { SIDEBAR_ICON_HEIGHT } from '../layout/Sidebar'
 
 export function SectionChoice({ label, selected, icon, onClick }) {
   return (
